+++ src/icon/index.tsx
@@ -5,16 +5,21 @@ import CircleIcon from 'icon/circleIcon'
 import ComfortIcon from 'icon/comfortIcon'
 import CrossIcon from 'icon/crossIcon'
 import DecreaseIcon from 'icon/decreaseIcon'
+import DoubleArrowIcon from 'icon/doubleArrowIcon'
 import EyeIcon from 'icon/eyeIcon'
 import FacebookIcon from 'icon/facebookIcon'
 import IncreaseIcon from 'icon/increaseIcon'
+import InstagramIcon from 'icon/instagramIcon'
 import LadyIcon from 'icon/ladyIcon'
 import LightningIcon from 'icon/lightningIcon'
 import PinIcon from 'icon/pinIcon'
 import QuestionIcon from 'icon/questionIcon'
-import SearchIcon from 'icon/searchIcon'
 import StarIcon from 'icon/starIcon'
+import SearchIcon from 'icon/searchIcon'
+import TrashIcon from 'icon/trashIcon'
+import TwitterIcon from 'icon/twitterIcon'
 import VkontakteIcon from 'icon/vkontakteIcon'
+import YoutubeIcon from 'icon/youtubeIcon'
 import ClockIcon from 'icon/clockIcon'
 import CrosshairIcon from 'icon/crosshairIcon'
 import ProximityIcon from 'icon/proximityIcon'
@@ -29,9 +34,11 @@ export {
   CrossIcon,
   CrosshairIcon,
   DecreaseIcon,
+  DoubleArrowIcon,
   EyeIcon,
   FacebookIcon,
   IncreaseIcon,
+  InstagramIcon,
   LadyIcon,
   LightningIcon,
   PinIcon,
@@ -39,5 +46,8 @@ export {
   QuestionIcon,
   SearchIcon,
   StarIcon,
+  TrashIcon,
+  TwitterIcon,
   VkontakteIcon,
+  YoutubeIcon,
 }